backend/functions/telemetry_firestore_trigger/requirements.txt
google-cloud-firestore==2.13.1
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
functions-framework==3.5.0
google-events==0.13.0
protobuf>=3.19.5,<5.0.0